Add optional 'IsReference' parameter to [AsyncTestSuite].
When true, the test suite is loaded from the assembly where the 'Type'
is declared in, not current one containing the [AsyncTestSuite] attribute.

We can now omit the "--dependency=Xamarin.WebTests.Console.dll" argument on
the command-line and only specify "Xamarin.WebTests.Console.dll" as test
suite; it will automatically register the dependencies and load the real
test suite from "Xamarin.WebTests.dll".