The big Xamarin.WebTests.HttpClientTests / HttpRequestTests refactoring.
Summary of changes:
===================

* Added new Xamarin.WebTests.HttpClientTests and HttpRequestTests namespaces
  in the Xamarin.WebTests.Tests project.

* HttpClientTestType and HttpRequestTestType is now gone - for each of the
  previous enum values, a new class has been created in the above mentioned
  new namespaces.

  This also moved the test logic itself from Xamarin.WebTests.Framework into
  Xamarin.WebTests.Tests.

* HttpClientTestRunner and HttpRequestTestRunner now calls back into these
  new test fixtures via a newly introduced interface.

* The same will be done with HttpInstrumentationTestType in near future.

Rationale behind this:
======================

When I wrote the first version of the HttpInstrumentationTestRunner, the
business logic behind it was large and complex compared to the actual
test code.  We were essentially sending a request to the internal server
and reading a response back from it - but it had to be setup in a very
peculiar manner.

While this test runner was meant to quickly test a specific issue that I
had back then, I quickly fell in love with its underlying concept and kept
adding more and more tests to it, so it grew bigger and bigger over time.

Eventually, it was getting so big that it had to be split into smaller
pieces - and HttpClientTestRunner and HttpRequestTestRunner were born.
As time passed by, these also became too big, so yet again they were
split into even smaller pieces - spanning about ten different classes
in the end.

Unfortunately, this made it much harder to read and understand the
underlying test logic as you had to look in up to three different source
files for each particular test.

Finally, the idea was born to get rid of all those enums and put each
test into its own test fixture.  However, to do this nicely, the underlying
Xamarin.AsyncTests framework needed a few improvements.  This was something
I've always been scared of doing, but I figured that having a dedicated
test suite for the Xamarin.AsyncTests framework would mostly mitigate the
risk and also make things much more stable and reliable going forward.

The Xamarin.WebTests.HttpRequestTests namespace was the first one that I
did and when you look at the HttpClientTests one you'll see that I got a
few ideas while working on it that I then implemented in the framework.

Adding new tests is now really easy - instead of adding a new enum value
and then editing three or even four different source files (and getting
runtime errors when you forgot any one of them), you now simply create a
new class and implement a few methods in it.

There are currently 19 tests in the HttpClientTests namespace and 27 in
HttpRequestTests.  Some of these are parameterized such as for instance
HttpClientTests.ReuseConnection and others.