Experimental "forked" support =============================
This is not entirely complete yet and might break non-console platforms.

This is another set of larger framework changes, this time adding new features which will
be used by a set of new tests that I indend on writing in near future.

The existing [Forked] support has been extended to support cross-appdomain
and cross-project forks.

You can now write a test like this

		[AsyncTest]
		public static void Test (
			TestContext ctx, MyHost host,
			[Fork (ForkType.Process)] IFork fork)
		{
			ctx.LogMessage ($"Martin Test: {ctx.FriendlyName} {host} {fork}");
			ctx.Assert (host.IsForked);
		}

and it will spawn an external process (or appdomain with `ForkType.Domain`) and
serialize the `MyHost` instance accross process / appdomain boundaries.

This serialization is something that's already being used on all non-console
platforms since pretty much the beginning of time - the only new thing is that
tests now have access to it via a new `IForkedTestInstance` interface.

We're essentially following existing scope / lifecyle rules here - everything
that comes after the `IFork` argument will only ever exist on the remote side,
anything that comes before it needs to be serialized either via the existing
`ITestParameter` mechanism or the new `IForkedTestInstance` interface.

The idea is that you should be able to essentially split the client / server
side of a test across process / appdomain boundaries.

For instance, you could use the internal web server to listen for requests,
then fork off an external process to connect to it - which is great for stress /
performance tests.

When implementing `IForkedTestInstance`, either its `Serialize()` or its
`Deserialize()` method will be called - these methods are guaranteed to be
called exactly once and this will happen prior to `ITestInstance.Initialize()`.